+++ packages/core/config/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @executor-js/config
 
+## 1.5.35
+
+### Patch Changes
+
+- Updated dependencies [[`1b9b1f1`](https://github.com/UsefulSoftwareCo/executor/commit/1b9b1f10313834a625a411169ebf83f6181589df)]:
+  - @executor-js/sdk@1.5.35
+
 ## 1.5.34
 
 ### Patch Changes